+++ R/errors.R
@@ -228,7 +228,7 @@ trainingaccuracy <- function(f, test, d, D) {
 #' \bold{22}(4), 679-688. Hyndman, R.J. and Athanasopoulos, G. (2018)
 #' "Forecasting: principles and practice", 2nd ed., OTexts, Melbourne, Australia.
 #' Section 3.4 "Evaluating forecast accuracy".
-#' \url{https://otexts.org/fpp2/accuracy.html}.
+#' \url{https://otexts.com/fpp2/accuracy.html}.
 #' @keywords ts
 #' @examples
 #'
@@ -254,7 +254,7 @@ accuracy.default <- function(object, x, test = NULL, d = NULL, D = NULL, f = NUL
     object <- f
   }
   if (!any(is.element(class(object), c(
-    "mforecast", "forecast", "ts", "integer", "numeric",
+    "ARFIMA", "mforecast", "forecast", "ts", "integer", "numeric",
     "Arima", "ets", "lm", "bats", "tbats", "nnetar", "stlm", "baggedModel"
   )))) {
     stop("First argument should be a forecast object or a time series.")