+++ NEWS.md
@@ -1,4 +1,10 @@
-## forecast 8.11 (21 March 2020)
+# forecast (development version)
+  * Fixed forecasts from Arima with drift with initial NAs.
+  * Fixed season colours in `gglagplot()` to match y-axis (original data).
+  * Fixed facet order for classical decomposition `autoplot()`
+  * Fixed `summary()` erroring for `tslm()` models containing NA values.
+
+## forecast 8.12 (21 March 2020)
   * Fixed bias adjusted forecast mean for ARIMA forecasts.
   * Improved naming of `accuracy()` generic formals.
   * Fix seasonal periods for `taylor` dataset.
@@ -9,7 +15,7 @@
 ## forecast 8.10 (4 December 2019)
   * Updates to remove new CRAN errors
   * Bug fixes
-  
+
 ## forecast 8.9 (22 August 2019)
   * Updates for CRAN policies on Suggests packages
   * Bug fixes